형변환 ㅅ주어

+++ util.py
@@ -106,7 +106,7 @@ def train(train_loader, model, loss_func, device, optimizer, scheduler=None):
     with tqdm.tqdm(train_loader, total=len(train_loader), desc="Train", file=sys.stdout) as iterator:
         for train_x, train_y in iterator:
             train_x = train_x.float().to(device)
-            train_y = train_y.float().to(device)
+            train_y = train_y.long().to(device)
             output = model(train_x)            
             loss = loss_func(output, train_y)
             
@@ -140,7 +140,7 @@ def validate(valid_loader, model, loss_func, device, scheduler=None):
     with tqdm.tqdm(valid_loader, total=len(valid_loader), desc="Valid", file=sys.stdout) as iterator:
         for train_x, train_y in iterator:
             train_x = train_x.float().to(device)
-            train_y = train_y.float().to(device)
+            train_y = train_y.long().to(device)
 
             with torch.no_grad():
                 output = model(train_x)
